version presque final

+++ backend/database/migrations/2025_12_15_185838_remove_nbexemplaire_from_attestationscolarite_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('attestationscolarite', function (Blueprint $table) {
-            $table->dropColumn('nbrExemplaire');
+            if (Schema::hasColumn('attestationscolarite', 'nbrExemplaire')) {
+                $table->dropColumn('nbrExemplaire');
+            }
         });
     }
